Large refactoring for introducing data abstraction
