+++ src/ctDNAtool/preprocessors/find_tss.py
@@ -1,9 +1,5 @@
 import attr
-import re
 import csv
-import py2bit
-from enum import Enum, auto
-from natsort import natsorted
 
 from .transcript_annotation import (
     pull_tx_id,
@@ -12,6 +8,7 @@ from .transcript_annotation import (
     pull_tx_type,
 )
 from .tss import TranscriptionStartSite, get_header
+from .utils import is_autosome
 
 
 @attr.s
@@ -21,14 +18,6 @@ class Tx_annotation:
     end = attr.ib()
     strand = attr.ib()
     additionals = attr.ib()
-
-
-def is_autosome(chrom):
-    return re.match(r"chr[3-9]$|chr1[0-9]?$|chr2[0-2]?$", chrom) is not None
-
-
-def is_autosome_or_x(chrom):
-    return re.match(r"chr[3-9]$|chr1[0-9]?$|chr2[0-2]?$|chrX$", chrom) is not None
 
 
 def get_transcript_annotations(file_path):
@@ -66,64 +55,7 @@ def determine_TSS_and_format_data(tx_anno):
     )
 
 
-class Chromosomes(Enum):
-    AUTOSOMES = auto()
-    AUTOSOMES_X = auto()
-
-
-def select_chrom_test(chroms):
-    CHROM_TESTS = {
-        Chromosomes.AUTOSOMES: is_autosome,
-        Chromosomes.AUTOSOMES_X: is_autosome_or_x,
-    }
-    return CHROM_TESTS.get(chroms)
-
-
-def preprocess_bin_genome_Mbp(
-    genome_ref_file, output_file, mbp=1.0, chromosomes=Chromosomes.AUTOSOMES_X
-):
-    """This function will given a genome reference file in .2bit format,
-    create a bed file splitting the genome in bins of size mbp.
-
-    If the last bin of a chromosome is smaller than the given mbp,
-    the bin will be thrown away
-
-    :param genome_ref_file: File path to a .2bit file
-    :type genome_ref_file: str
-    :param output_file: The path to where the outputting bed file
-                        is stored.
-    :type output_file: str
-    :param mbp: Bin size in Mbp.
-    :type mbp:
-    :param chromosomes: Choose wich chromosomes to include in bed file.
-    :type Chromosomes:
-    """
-    # NOTE: If no output file path is given, the output should prob. be
-    #       outputtet on stdout. Nice for piping.
-    tb = py2bit.open(genome_ref_file)
-    bin_size = int(mbp * 10 ** 6)
-    try:
-        chroms = tb.chroms()
-    finally:
-        tb.close()
-    with open(output_file, "w") as fp:
-        bed_writer = csv.writer(fp, delimiter="\t")
-        bed_writer.writerow(["#chrom", "start", "end", "name", "score", "strand"])
-        for chr_name in natsorted(
-            filter(chroms.keys(), key=select_chrom_test(chromosomes))
-        ):
-            length = int(chroms[chr_name])
-            pos_pairs = zip(
-                range(0, length, bin_size), range(bin_size, length, bin_size)
-            )
-            for start, end in pos_pairs:
-                bed_writer.writerow(
-                    [chr_name, start, end, "{}_{}".format(chr_name, start), 0, "+"]
-                )
-    return output_file
-
-
-def preprocess(input_file, region_size, bed_file, tss_file):
+def find_tss(input_file, region_size, bed_file, tss_file):
     """This function will given a gencode annotation file, find all transcripts
     and determine the Transcription Start Site (TSS) for the transcript.
     Information about the TSS will be stored in the tss file with metadata
